Work continues, added global consumer list
Well, so far the idea is to have vectors of functions that will
fullfill various tasks. Currently there are *row consumers*, that
get to process every xlogfile row as it is parsed; and *global consumers*
get invoked after the xlogfile read in is complete.

These functions will together build the scoreboard structure, that
would be then fed into templates.