Implement Best Of 13 trophy
This compiles per-player "best13" subkey that contains: "cnt" - count of
player's "best of 13" games, "games" which lists the games that
contribute to that result and "when" which give the endtime of the last
of the games (giving the time when the result was achieved).

In %s{trophies}{best13} there is simple list of players orederd by their
Best Of 13 achievement.  Ties are broken by who got there first.