Implement two ways of displaying death/conduct fields
The two modes are controlled by 'aswon' parameter passed to game.tt
template.  If it evaluates as true, the death field is not displayed and
conducts are displayed as two TD's where the first shows number of
conducts and the second spells out the conducts in human readable form.
If 'aswon' evaluates as false, then normal death reason is shown and
in case of ascending game, conducts are appended in human readable form.

This change will make showing tables of ascending games little more
efficient and readable.