Change the way static info about trophies is handled
This moves the static info (lists of trophies, orderings, display names)
from the code to the configuration file.  To make this available to the
templates, the config data are included in the data passed to them.  The
'unique' branch of the config tree is removed because it contains the
compiled regexes objects which interfere with JSON export.