Implement scummed games detection and hiding
The scoreboard now detects games that are scummed by criteria configured
in the config file:

scum.minturns - any game with less than this amount of turns is always
considered scummed

scum.minquitturns - any game that has death reason "quit" or "escaped" and
has less than this amount of turns is considered scummed

--

This commit implements:

* every game is tagged with '_scum' field with values 0 or 1, according
  to the configured criteria

* every clan game is also tagged with '_cid' - a sequence number of all
  games for given clan; this must be done so that we can show this
  numbering on the clan page

* for individual player pages, all non-scummed games are listed; if the
  number of games is small than 10, all games are listed (to avoid some
  players having games but showing none)

* for clan pages and front page, only non-scummed games are shown

Note, that the above only affects how the games are _shown_ on the scoreboard,
nothing else. Even scummed games can therefore be shown on trophy pages when
ascension is not condition of getting a trophy; scummed games also still
break the streaks required for the "with the bells on" trophies.

Also note, that the sequence numbers shown it tables that list played games
can be discontinous, because scummed games are hidden.