Add a 'when' field to entries in player score breakdown
Per-player Best In Show score breakdown has now time reference
in the fourth slot in the descriptor array. This applies for all
trophies except for Recognition Trophies, which will be (hopefully)
implemented with the next commit. This is intended as a way to
break ties in the Best In Show ladder.